updating routes.js to deal with issues related to loading in node vs webpack environmnets. Updating js components based on comments from pull request.  Updated PR template to add additional item to regenerate routes file when the underlying rails routes have been modified
