+++ .eslintignore
@@ -1,3 +1,3 @@
 dist/**/*
 test/test_helper.js
-webpack/routes.js
+webpack/_routes.js
