+++ webpack/components/QuestionWidget.js
@@ -1,6 +1,6 @@
 import React, { Component, PropTypes } from 'react';
-var routes = require("../routes.js");
-if(routes.Routes){ routes=routes.Routes;}
+import Routes from "../routes.js";
+
 export default class QuestionWidget extends Component {
   constructor(props) {
     super(props);
@@ -13,7 +13,7 @@ export default class QuestionWidget extends Component {
               <div className="question-container">
 
                   <ul className="list-inline">
-                      <li><a href={routes.question_path(this.props.question)}>{this.props.question.content}</a></li>
+                      <li><a href={Routes.question_path(this.props.question)}>{this.props.question.content}</a></li>
                       <li className="pull-right">
                           <a>
                               <span className="glyphicon glyphicon-signal"></span>
@@ -33,13 +33,13 @@ export default class QuestionWidget extends Component {
                               </a>
                               <ul className="dropdown-menu">
                               <li>
-                                <a href={routes.revise_question_path(this.props.question)}>Revise</a>
+                                <a href={Routes.revise_question_path(this.props.question)}>Revise</a>
                               </li>
                               <li>
-                                <a href={routes.question_path(this.props.question)}>Details</a>
+                                <a href={Routes.question_path(this.props.question)}>Details</a>
                               </li>
                                   <li>
-                                  <a data-confirm="Are you sure?" rel="nofollow" data-method="delete" href={routes.question_path(this.props.question)}>Delete</a>
+                                  <a data-confirm="Are you sure?" rel="nofollow" data-method="delete" href={Routes.question_path(this.props.question)}>Delete</a>
                                   </li>
                               </ul>
                           </div>
